[피드백 반영] Const 파일 분리 1. Storyboard 2. ViewController
