+++ Instagram/Instagram/Constants/Const.swift
@@ -1,0 +1,16 @@
+//
+//  Const.swift
+//  Instagram
+//
+//  Created by 김영인 on 2022/04/12.
+//
+
+import Foundation
+
+struct Const {
+    /*
+    Const.Storyboard.Name.login 과 같이 사용.
+    Const.ViewController.Identifier.loginVC 과 같이 사용.
+    */
+}
+
