+++ Instagram/Instagram/Constants/Storyboard.swift
@@ -1,0 +1,21 @@
+//
+//  Storyboard.swift
+//  Instagram
+//
+//  Created by 김영인 on 2022/04/12.
+//
+
+import Foundation
+
+extension Const {
+    // Storyboard 의 name 을 상수로 관리합니다.
+    struct Storyboard {
+        struct Name {
+            static let main = "Main"
+            static let login = "Login"
+            static let signupName = "SignupName"
+            static let signupPassword = "SignupPassword"
+            static let success = "Success"
+        }
+    }
+}
